live data with local server

+++ app/src/main/java/com/axelliant/hrms/model/attendance/AttendanceDetail.kt
@@ -12,5 +12,5 @@ data class AttendanceDetail(
     val shift: String,
     val shift_timings: String,
     val status: String,
-    val working_hours: Int
+    val working_hours: Double
 )
