+++ app/src/main/java/com/axelliant/hrms/model/leave/SelfLeaveStats.kt
@@ -2,11 +2,11 @@ package com.axelliant.hrms.model.leave
 
 class SelfLeaveStats {
 
-    val self_approved: Int=0
-    val self_pending: Int=0
-    val self_reject: Int=0
-    val remaining_leave: Int=0
-    val total_leave: Int=0
+    val self_approved: Double=0.0
+    val self_pending: Double=0.0
+    val self_reject: Double=0.0
+    val remaining_leave: Double=0.0
+    val total_leave: Double=0.0
 
 
 }
