+++ app/src/main/java/com/axelliant/hrms/model/leave/TeamLeaveStats.kt
@@ -1,11 +1,11 @@
 package com.axelliant.hrms.model.leave
 
 class TeamLeaveStats{
-    val total_team_members: Int=0
-    val team_approved: Int=0
-    val team_pending: Int=0
-    val team_reject: Int=0
-    val all_leaves: Int=0
+    val total_team_members: Double=0.0
+    val team_approved: Double=0.0
+    val team_pending: Double=0.0
+    val team_reject: Double=0.0
+    val all_leaves: Double=0.0
 
 
 }
